feat(Query scopesSearch): Search scopes, order them and pages.
When it searches for scopes, it returns all the results that match the filters sent.
If no results are found, it returns an empty array.
If sorting criteria are submitted, it returns the sorted results.
If no paging criteria are sent, the results are paged with the default paging (See PAGINATION_DEFAULT in src/shared/queries.constant.ts).